requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64

numpy~=1.20.3
opensimplex~=0.3
matplotlib~=3.4.2
Pillow~=8.2.0
PyYAML~=5.4.1